# src/nf_server/swagger.yml
info:
  description: Swagger schema for Nextflow server
  title: nf-server
  version: 1.0.0
security:
  - APIKeyHeader: []
paths:

  /ping:
    get:
      tags:
      - ping
      summary: checks whether Nextflow is running on the server
      operationId: pingNextflowServer
      responses:
        "200":
          description: Server is up and running
          schema:
            $ref: "#/definitions/error"
        "401":
          $ref: "#/responses/UnauthorizedError"
        "503":
          description: Nextflow is not running on the server
          schema:
            $ref: "#/definitions/error"

  /submit:
    post:
      tags:
      - submit
      operationId: submitWorkflow
      parameters:
      - in: "body"
        name: "body"
        description: "Submit a workflow"
        schema:
          $ref: "#/definitions/workflow_request"
      responses:
        "202":
          description: Workflow accepted for processing successfully
          schema:
            type: object
            required:
            - workflow_id
            properties:
              workflow_id:
                type: string
        "401":
          $ref: "#/responses/UnauthorizedError"
        default:
          description: generic error response
          schema:
            $ref: "#/definitions/error"
        "422":
          description: Workflow was not accepted for processing
          schema:
            $ref: "#/definitions/error"

  /status/{workflow_id}:
    get:
      tags:
      - status
      summary: checks workflow status
      operationId: checkWorkflowStatus
      parameters:
      - name: workflow_id
        in: path
        type: string
        required: true
      responses:
        "200":
          description: Successful operation
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
        "401":
          $ref: "#/responses/UnauthorizedError"
        default:
          description: generic error response
          schema:
            $ref: "#/definitions/error"
produces:
- application/io.goswagger.examples.casbin.v1+json
schemes:
- http
- https
swagger: "2.0"
definitions:

  workflow_request:
    type: object
    required:
    - workflow
    properties:
      workflow:
        type: string
      wf_params:
        type: object
      nf_params:
        type: object
      file_inputs:
        type: object

  error:
    type: object
    required:
    - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
securityDefinitions:
   APIKeyHeader:
     type: apiKey
     in: header
     name: X-API-Key
responses:
  UnauthorizedError:
    description: API key is missing or invalid
    headers:
      WWW_Authenticate:
        type: string